dcap: Add stubs to build kbs without installing DCAP
**THIS IS A HACK!!!**, but this is a needed hack in order to build kbs
on an offline environment, where Intel DCAP packages cannot be
installed.

sgx_dcap_quoteverify_stubs brings in all the headers needed to build
everything we need from kbs side, and it also brings a very simple .cpp
file that is just a stub, returns an error, nothing else than that.

The expected usage of this is:
* A downstream would using such stubs
* During Trustee deployment, Intel DCAP libraries would be downloaded,
  installed into the kbs container, becoming then available to be used
  during runtime.

As already mentioned, this is not a proper solution, the proper solution
would be using the DCAP packages provided either by Intel (if possible),
or by the distros (when that's the requirement).

A new Dockerfile.fidencio was added, so whoever ends up using this, can
confirm that the project can be built, and also check how to build the
sgx_dcap_quoteverify_stubs itself.

Signed-off-by: Fabiano Fidêncio <fabiano.fidencio@intel.com>